click on moroword now show the sentence id
When we render the dictionary page for the first time, we also create a global_click_database which contains the id of each sentence and the moroword+definition pair. When a moroword is click, we then find occurrence using the global_click_database. Right now the side tab will show the sentence id where that moroword+definition pair occurred. The dictionary page loading is slower than before mainly due to making every moroword clickable. (also a little slower due to creating global_click_database but shouldnt be that much compare to rendering the clickable links). Each moroword, when clicked,  will take couple seconds to load all the occurrence. I try to speed this up as much as possible by preprocessing the json file (splitting, replacing, breaking out -, find verb etc.) and keeping it in memory