main page + formatting
I added content to the main page, mostly just to practice using
Semantic UI a bit. I also changed the main menu to a floating menu that
will stay at the top of the screen, changed the name of dictionary to
concordance, switched the order of the text and the dictionary page,
and embedded the whole page in a ui text container which is why it now
has the nicer look.