Make just one schwa character in the database
Fix a bug where two schwas were appearing. Replaced all unicode
characters \u0259 (Latin Small Letter Schwa) with \u01DD (Latin Small
Letter Turned E).

Chose the turned e over the schwa because it looked a bit more congruent
with the font of the rest of the charactes, but this is arbitrary.
Changing the order of the two '\u' arguments in the .replace function
would reverse this.