Updated static sentences and story files.
Cloudant is soon to discontinue their offering that we were using to
proxy the data from lingsync. Rather than attempt to migrate that
functionality to a new location, just move the data to static files in
the github repo.

If any modifications need to be done to the data, the JSON files can be
edited directly.